requirements.txt
requests
aiogram
python-dotenv